Disable auto-remembering of thread/stack referenced old objects for Concurrent Scavenger. This optimization is not safe since it may leave such references to Evacuate space during concurrent copy-phase phase.
