+++ example/glue/ScavengerRootScanner.hpp
@@ -88,8 +88,10 @@ public:
 			env->_currentTask->releaseSynchronizedGCThreads(env);
 		}
 	}
-
+	
+#if !defined(OMR_GC_CONCURRENT_SCAVENGER)
 	void rescanThreadSlots(MM_EnvironmentStandard *env) { }
+#endif
 
 	void scanClearable(MM_EnvironmentBase *env)
 	{
